Fix injection of dependencies
