--- app/src/main/java/tarsila/costalonga/notasapp/di/ModuleDatabase.kt
@@ -1,31 +1,0 @@
-package tarsila.costalonga.notasapp.di
-
-import android.content.Context
-import androidx.room.Room
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import tarsila.costalonga.notasapp.data.local.NotasDao
-import tarsila.costalonga.notasapp.data.local.NotasRoom
-
-@InstallIn(SingletonComponent::class)
-@Module
-object ModuleDatabase {
-    @Provides
-    fun proverDatabase(
-        @ApplicationContext context: Context,
-    ): NotasRoom {
-        return Room.databaseBuilder(
-            context.applicationContext,
-            NotasRoom::class.java,
-            "notas_bd",
-        ).fallbackToDestructiveMigration().build()
-    }
-
-    @Provides
-    fun proverDao(dtBase: NotasRoom): NotasDao {
-        return dtBase.notasDao
-    }
-}
